🎨 ui/index.ts 관심사 분리
- 일단 나누고 생각해보기 위해 러프하게 나눔.
