+++ src/shared/ui/Select/Select.tsx
@@ -1,0 +1,6 @@
+import * as SelectPrimitive from '@radix-ui/react-select';
+// 선택 컴포넌트
+
+export const Select = SelectPrimitive.Root;
+export const SelectGroup = SelectPrimitive.Group;
+export const SelectValue = SelectPrimitive.Value;
