+++ src/shared/ui/Select/SelectContent.tsx
@@ -1,0 +1,24 @@
+import { forwardRef } from 'react';
+import * as SelectPrimitive from '@radix-ui/react-select';
+
+interface SelectOptionsProps
+  extends React.HTMLAttributes<HTMLDivElement>,
+    SelectPrimitive.SelectContentProps {
+  className?: string;
+}
+
+export const SelectContent = forwardRef<HTMLDivElement, SelectOptionsProps>(
+  ({ className, children, position = 'popper', ...props }, ref) => (
+    <SelectPrimitive.Portal>
+      <SelectPrimitive.Content
+        ref={ref}
+        className={`relative z-50 min-w-[8rem] overflow-hidden rounded-md border bg-white text-popover-foreground shadow-md data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 data-[side=bottom]:slide-in-from-top-2 data-[side=left]:slide-in-from-right-2 data-[side=right]:slide-in-from-left-2 data-[side=top]:slide-in-from-bottom-2 ${className}`}
+        position={position}
+        {...props}
+      >
+        <SelectPrimitive.Viewport className='p-1'>{children}</SelectPrimitive.Viewport>
+      </SelectPrimitive.Content>
+    </SelectPrimitive.Portal>
+  ),
+);
+SelectContent.displayName = SelectPrimitive.Content.displayName;
