+++ src/shared/ui/Table/TableBody.tsx
@@ -1,0 +1,12 @@
+import { forwardRef } from 'react';
+
+interface TableBodyProps extends React.HTMLAttributes<HTMLTableSectionElement> {
+  className?: string;
+}
+
+export const TableBody = forwardRef<HTMLTableSectionElement, TableBodyProps>(
+  ({ className, ...props }, ref) => (
+    <tbody ref={ref} className={`[&_tr:last-child]:border-0 ${className}`} {...props} />
+  ),
+);
+TableBody.displayName = 'TableBody';
